Refactor stringset handling to dedicated file
